% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta.R
\name{get_beta}
\alias{get_beta}
\title{get_beta}
\usage{
get_beta(
  ticker,
  index,
  star_date = Sys.Date() - 365 * 5,
  end_date = Sys.Date()
)
}
\arguments{
\item{ticker}{stock ticker from yahoo as a string. Example: 'EQNR.OL'}

\item{index}{index ticker from yahoo as a string. Example: 'OSEBX.OL'}

\item{end_date}{end date. Default is today}

\item{start_date}{starting date. Default is 5 years from today}
}
\value{
the estimated beta
}
\description{
estimate beta for a stock using data from yahoo finance
}
\examples{
get_beta('EQNR.OL', 'OSEBX.OL')
get_beta('EQNR.OL', 'OSEBX.OL', '2021-01-01', '2023-01-01')

}
