% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta.R
\name{get_beta}
\alias{get_beta}
\title{get_beta}
\usage{
get_beta(
  ticker,
  index,
  start_date = Sys.Date() - 365 * 5,
  end_date = Sys.Date()
)
}
\arguments{
\item{ticker}{stock ticker from yahoo as a string}

\item{index}{index ticker from yahoo as a string}

\item{end_date}{}
}
\value{
Returns the estimated beta
}
\description{
estimate beta for a stock using data from yahoo finance
}
\examples{
get_beta('EQNR.OL', 'OSEBX.OL')
get_beta('EQNR.OL', 'OSEBX.OL', '2015-01-01', '2017-01-01')

}
