% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_beta.R
\name{get_beta}
\alias{get_beta}
\title{get_beta}
\usage{
get_beta(
  ticker,
  index,
  start_date = Sys.Date() \%>\% floor_date("month") \%m-\% months(61),
  end_date = Sys.Date() \%>\% floor_date("month")
)
}
\arguments{
\item{ticker}{stock ticker from yahoo as a string}

\item{index}{index ticker from yahoo as a string}

\item{start_date}{start of estimation period}

\item{end_date}{end of estimation period}
}
\value{
Returns the estimated beta
}
\description{
Estimates the beta for a stock using data from yahoo finance. By default it uses 5 years of monthly data. The same as yahoo finance
}
\examples{
get_beta('TSLA', '^GSPC')
get_beta('EQNR.OL', 'OSEBX.OL', '2015-01-01', '2017-01-01')

}
